refactor: widen smm_search_s.sweep_width to uint64_t
The public getter returns uint64_t, but the field was uint32_t, requiring an
implicit widening conversion every read. Server-side storage is signed 32-bit,
so no values are lost in practice, but matching the public type removes the
mismatch.

Fixes todo/claude/04-sweep-width-type-mismatch.md

+++ src/smm-asset-internal.h
@@ -83,7 +83,7 @@ struct smm_search_s
 	char *url;
 	uint64_t distance;
 	uint64_t length;
-	uint32_t sweep_width;
+	uint64_t sweep_width;
 };
 
 struct smm_curl_res_s
